pull out memstore logic
